114/06/07_Build the UI and send data via API
Completely convert the original backend logic that was using Flask to FastAPI, and spend time learning how to process data through API logic in the frontend with React. Set up a basic React front-end for users to upload images or videos, and connect the data to the API; plan for future image processing logic.